# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: CloudCarbon API
  description: |
    Open-source multi-cloud GreenOps platform for unified cost and carbon optimisation.

    **Authentication**: All endpoints (except `/health`) require either:
    - A `Bearer` JWT access token in the `Authorization` header, or
    - A `cc_`-prefixed API key in the `X-API-Key` header.

    **Session status**: Session 5 (final) — agent infrastructure, five autonomous agents,
    notification delivery, Prometheus metrics, structured logging, rate limiting, and
    production Helm chart are complete.
  version: "0.5.0"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.cloudcarbon.io
    description: Production (future)

# =============================================================================
# Security
# =============================================================================
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# =============================================================================
# Tags
# =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key prefixed with `cc_`, obtained from the dashboard

  # ---------------------------------------------------------------------------
  # Shared Parameters
  # ---------------------------------------------------------------------------
...
  schemas:
...
    CloudAccountCreate:
      type: object
      required: [name, provider, account_identifier]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Human-readable account name
        provider:
          type: string
          enum: [aws, azure, gcp, alibaba]
        account_identifier:
          type: string
          description: Provider account ID (AWS account ID, Azure subscription ID, GCP project ID)
        config:
          type: object
          additionalProperties: true
          description: |
            Provider-specific configuration:
            - AWS: `s3_bucket`, `s3_prefix`, `report_name`, `aws_region`
            - Azure: `subscription_id`, `resource_group`
            - GCP: `project_id`, `dataset_id`, `table_id`
            - Alibaba: `region_id`, `billing_cycle`
        credentials:
          type: object
          additionalProperties:
            type: string
          description: |
            Provider credentials (stored securely as credentials_ref, never returned):
            - AWS: `aws_access_key_id`, `aws_secret_access_key`
            - Azure: `azure_tenant_id`, `azure_client_id`, `azure_client_secret`
            - GCP: `service_account_json`
            - Alibaba: `access_key_id`, `access_key_secret`
...
    CloudAccountListResponse:
      type: object
      required: [accounts, total]
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/CloudAccountResponse'
        total:
          type: integer

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: List cloud accounts
      description: |
        Return all active cloud accounts for the current tenant.
        Includes `last_sync_at` timestamp and `record_count` for each account.
        Requires `viewer` role or higher.
      operationId: listAccounts
      responses:
        "200":
          description: List of cloud accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudAccountListResponse'
        "401":
          description: Not authenticated

    post:
      tags: [Accounts]
      summary: Create a cloud account connection
      description: |
        Create a new cloud account connection.
        Validates credentials by making a test API call to the provider before saving.
        Raw credentials are never stored; only a `credentials_ref` is kept.
        Requires `admin` role.
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudAccountCreate'
            examples:
              aws:
                summary: AWS account with S3 CUR
                value:
                  name: "Production AWS"
                  provider: aws
                  account_identifier: "123456789012"
                  config:
                    s3_bucket: "my-cur-bucket"
                    s3_prefix: "cur/reports"
                    report_name: "cloudcarbon-cur"
                    aws_region: "us-east-1"
                  credentials:
                    aws_access_key_id: "AKIAIOSFODNN7EXAMPLE"
                    aws_secret_access_key: "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
              azure:
                summary: Azure subscription
                value:
                  name: "Production Azure"
                  provider: azure
                  account_identifier: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                  config:
                    subscription_id: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                  credentials:
                    azure_tenant_id: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                    azure_client_id: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                    azure_client_secret: "your-client-secret"
      responses:
        "201":
          description: Cloud account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudAccountResponse'
        "401":
          description: Not authenticated
        "403":
          description: Insufficient role (requires admin)
        "422":
          description: Credential validation failed or invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{account_id}:
    delete:
      tags: [Accounts]
      summary: Delete a cloud account
      description: |
        Soft-delete a cloud account (sets `status = 'inactive'`).
        Writes an audit log entry.
        Requires `admin` role.
      operationId: deleteAccount
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        "204":
          description: Account deleted
        "401":
          description: Not authenticated
        "403":
          description: Insufficient role (requires admin)
        "404":
          description: Account not found

  # ---------------------------------------------------------------------------
  # Ingestion
  # ---------------------------------------------------------------------------
  /ingest/upload:
    post:
      tags: [Ingestion]
      summary: Upload a billing file
      description: |
        Upload a billing file for ingestion. Supported formats:
        - **AWS CUR** (CSV, optionally gzip-compressed) — auto-detected from filename
        - **Azure Cost Management** export (CSV)
        - **GCP BigQuery** billing export (JSON)
        - **FOCUS 1.0** (CSV or Parquet) — validated against the FOCUS schema

        If the file contains validation errors, they are returned without ingesting any records.
        Requires `analyst` role or higher.
      operationId: uploadBillingFile
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Cloud account UUID to associate records with
        - name: provider
          in: query
          required: false
          schema:
            type: string
            enum: [aws, azure, gcp, alibaba]
          description: Provider hint. Auto-detected from filename if not provided.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Billing file (CSV, Parquet, or JSON)
      responses:
        "200":
          description: Ingestion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResult'
              examples:
                success:
                  summary: Successful ingestion
                  value:
                    inserted: 1250
                    updated: 48
                    skipped: 0
                    total: 1298
                    errors: []
                    validation_errors: []
                validation_errors:
                  summary: FOCUS validation errors
                  value:
                    inserted: 0
                    updated: 0
                    skipped: 0
                    total: 0
                    errors: ["3 validation error(s) found; no records ingested."]
                    validation_errors:
                      - row: 5
                        field: "EffectiveCost"
                        message: "Input should be a valid number"
                      - row: 12
                        field: "BillingPeriodStart"
                        message: "Input should be a valid datetime"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient role (requires analyst or higher)
        "404":
          description: Cloud account not found
  /ingest/sync/{account_id}:
    post:
      tags: [Ingestion]
      summary: Trigger async account sync
      description: |
        Start a background sync job for the specified cloud account.
        The sync fetches billing data directly from the provider API and ingests it.
        Returns a `job_id` immediately; poll `/ingest/sync/{job_id}/status` for progress.
        Requires `engineer` or `admin` role.
      operationId: triggerSync
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        "202":
          description: Sync job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobResponse'
              example:
                status: sync_started
                account_id: "550e8400-e29b-41d4-a716-446655440000"
                job_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient role (requires engineer or admin)
        "404":
          description: Cloud account not found
  /ingest/sync/{job_id}/status:
    get:
      tags: [Ingestion]
      summary: Get sync job status
      description: |
        Poll the status of an async sync job.
        Job status is stored in Redis with a 24-hour TTL.
      operationId: getSyncStatus
      parameters:
        - $ref: '#/components/parameters/JobIdPath'
      responses:
        "200":
          description: Sync job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobStatus'
              examples:
                running:
                  summary: Job in progress
                  value:
                    job_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                    account_id: "550e8400-e29b-41d4-a716-446655440000"
                    status: running
                    started_at: "2026-04-01T10:00:00Z"
                    completed_at: null
                    result: null
                    error: null
                completed:
                  summary: Job completed
                  value:
                    job_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                    account_id: "550e8400-e29b-41d4-a716-446655440000"
                    status: completed
                    started_at: "2026-04-01T10:00:00Z"
                    completed_at: "2026-04-01T10:02:34Z"
                    result:
                      inserted: 8420
                      updated: 312
                      skipped: 0
                      total: 8732
                      errors: []
                      validation_errors: []
                    error: null
        "401":
          description: Not authenticated
        "404":
          description: Job not found or expired

  # ---------------------------------------------------------------------------
  # Records (Session 3 will fully implement)
  # ---------------------------------------------------------------------------
...